requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
et-xmlfile==1.0.1
Flask==0.12.2
Flask-SQLAlchemy==2.2
idna==2.6
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.6
MarkupSafe==1.0
matplotlib==2.0.2
numpy==1.13.1
openpyxl==2.4.8
pandas==0.20.3
pylast==1.9.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
scipy==0.19.1
seaborn==0.8.1
six==1.11.0
spotipy==2.4.4
SQLAlchemy==1.1.14
urllib3==1.22
Werkzeug==0.12.2
psycopg2==2.4.4
